upgraded letters game -added DB to ligbdx
TODO:
-screen with choosing profile and proper updating of DB
-recognize if no records were added-then firstly insert users data
-2nd part: recegnition of speech+shown letter

+++ core/src/com/FallingLettersGame/game/FallingLettersGame.java
@@ -127,7 +127,7 @@ public class FallingLettersGame implements Screen {
 
         if ((generatedLetters>maxLetterNo) || (LettersCatched>maxLetterNo))
         {
-            game.setScreen(new ScoresScreen(game,LettersCatched));
+            game.setScreen(new ScoresScreen(game,(int)LettersCatched));
             dispose();
         }
